+++ lib/eva/client/remote_item.py
@@ -70,8 +70,10 @@ class RemoteUpdatableItem(eva.item.UpdatableItem):
         d = super().serialize(
             full=full, config=config, info=info, props=props, notify=notify)
         d['controller_id'] = self.controller.full_id
-        try: del d['config_changed']
-        except: pass
+        try:
+            del d['config_changed']
+        except:
+            pass
         return d
 
 
@@ -131,18 +133,36 @@ class RemoteUnit(RemoteUpdatableItem, eva.item.PhysicalItem):
         self.action_enabled = True
         self.status_labels = state.get('status_labels')
 
+    def update_nstate(self, nstatus=None, nvalue=None):
+        need_notify = False
+        if nstatus is not None:
+            try:
+                _s = int(nstatus)
+                if self.nstatus != _s:
+                    self.nstatus = _s
+                    need_notify = True
+            except:
+                logging.info('%s nstatus "%s" is not number, can not set' % \
+                        (self.full_id, nstatus))
+                eva.core.log_traceback()
+                return False
+        if nvalue is not None:
+            if nvalue == '': nv = 'null'
+            else: nv = nvalue
+            if self.nvalue != nv:
+                self.nvalue = nv
+                need_notify = True
+        if need_notify:
+            self.notify()
+        return True
+
     def mqtt_set_state(self, topic, data):
         super().mqtt_set_state(topic, data)
         try:
             if topic.endswith('/nstatus'):
-                try:
-                    self.nstatus = int(data)
-                    self.notify()
-                except:
-                    pass
+                self.update_nstate(nstatus=data)
             elif topic.endswith('/nvalue'):
-                self.nvalue = data
-                self.notify()
+                self.update_nstate(nvalue=data)
             elif topic.endswith('/action_enabled'):
                 self.action_enabled = eva.tools.val_to_boolean(data)
                 self.notify()
@@ -170,7 +190,7 @@ class RemoteMacro(eva.item.Item):
         self.controller = controller
         self.action_enabled = False
 
-    def update_config(self,cfg):
+    def update_config(self, cfg):
         super().update_config(cfg)
         if 'action_enabled' in cfg:
             self.action_enabled = cfg['action_enabled']